requirements/prod.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements/prod.in
#
asn1crypto==1.5.1
    # via scramp
cachetools==5.2.0
    # via google-auth
certifi==2022.6.15
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via flask
flask==2.1.2
    # via
    #   -r requirements/prod.in
    #   flask-wtf
flask-wtf==1.0.1
    # via -r requirements/prod.in
google-api-core==2.8.2
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.8.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.3.1
    # via google-cloud-storage
google-cloud-storage==2.4.0
    # via -r requirements/prod.in
google-crc32c==1.3.0
    # via google-resumable-media
google-resumable-media==2.3.3
    # via google-cloud-storage
googleapis-common-protos==1.56.3
    # via google-api-core
greenlet==1.1.2
    # via sqlalchemy
gunicorn==20.1.0
    # via -r requirements/prod.in
idna==3.3
    # via requests
importlib-metadata==4.11.4
    # via flask
itsdangerous==2.1.2
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.2
    # via flask
markupsafe==2.1.1
    # via
    #   jinja2
    #   wtforms
pg8000==1.29.1
    # via -r requirements/prod.in
protobuf==4.21.1
    # via
    #   google-api-core
    #   googleapis-common-protos
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
requests==2.28.0
    # via
    #   google-api-core
    #   google-cloud-storage
rsa==4.8
    # via google-auth
scramp==1.4.1
    # via pg8000
six==1.16.0
    # via google-auth
sqlalchemy==1.4.37
    # via -r requirements/prod.in
urllib3==1.26.9
    # via requests
werkzeug==2.1.2
    # via flask
wtforms==3.0.1
    # via flask-wtf
zipp==3.8.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
